Update Android gradle plugin to 4.1.0 version.
Move all android related code to separate file - AndroidPluginsApplier.kt.

Use new AGP apis and simplify internal code.
