+++ plugin/buildSrc/src/main/kotlin/PluginDependencies.kt
@@ -2,7 +2,7 @@ object PluginVersions {
     val kotlin = "1.4.10"
     val ktlintPlugin = "9.3.0"
     val gradlePublishPlugin = "0.12.0"
-    val androidPlugin = "3.6.3"
+    val androidPlugin = "4.1.0"
     val semver = "1.1.1"
     val jgit = "5.6.0.201912101111-r"
     val gradleWrapper = "6.6.1"
